==> 01-PreprocessCreateNewCSV-BP1-03.ipynb <==
Import libraries requred for the script

import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
from scipy.stats import norm
from sklearn.preprocessing import StandardScaler
from scipy import stats
import warnings
warnings.filterwarnings('ignore')
%matplotlib inline
from numpy.polynomial import polynomial as npp
from scipy.stats import linregress

Read the CSV files

df_pa = pd.read_csv('..\BridgestoneData\VMI_BP1.v1.csv')


Check the columns in the data

df_pa.describe()

df_pa.columns

Sort the data by timestamp

print(df_pa.shape)
df_pa_dupes = df_pa.duplicated(subset=['CUT_LENGTH', 'CONV_TIME', 'SPLICE_LENGTH_LEFT', 'SPLICE_LENGTH_MID', 
                                       'SPLICE_LENGTH_RIGHT', 'PRESSURE_S1_L', 'PRESSURE_S1_R', 'PRESSURE_S2_L',
                                       'PRESSURE_S2_R', 'PRESSURE_S3_L', 'PRESSURE_S3_R', 'PRESSURE_S4_L',
                                       'PRESSURE_S4_R', 'PRESSURE_S5_L', 'PRESSURE_S5_R', 'PRESSURE_S6_L',
                                       'PRESSURE_S6_R', 'PRESSURE_S7_L', 'PRESSURE_S7_R', 'PRESSURE_S8_L', 
                                       'PRESSURE_S8_R','RHT_SPLICE_SLOPE50', 'LFT_SPLICE_OK', 'MID_SPLICE_OK', 
                                       'RHT_SPLICE_OK', 'SPLICE_OK'])
df_pa = df_pa.drop_duplicates(subset=['CUT_LENGTH', 'CONV_TIME', 'SPLICE_LENGTH_LEFT', 'SPLICE_LENGTH_MID', 
                                      'SPLICE_LENGTH_RIGHT', 'PRESSURE_S1_L', 'PRESSURE_S1_R', 'PRESSURE_S2_L',
                                      'PRESSURE_S2_R', 'PRESSURE_S3_L', 'PRESSURE_S3_R', 'PRESSURE_S4_L',
                                      'PRESSURE_S4_R', 'PRESSURE_S5_L', 'PRESSURE_S5_R', 'PRESSURE_S6_L',
                                      'PRESSURE_S6_R', 'PRESSURE_S7_L', 'PRESSURE_S7_R', 'PRESSURE_S8_L', 
                                      'PRESSURE_S8_R','RHT_SPLICE_SLOPE50', 'LFT_SPLICE_OK', 'MID_SPLICE_OK', 
                                      'RHT_SPLICE_OK', 'SPLICE_OK'])
# df_pa = df_pa.drop_duplicates()
print(df_pa.shape)

print(df_pa.shape)
df_pa_dupes = df_pa.duplicated(subset=['CUT_LENGTH', 'CONV_TIME', 'SPLICE_LENGTH_LEFT', 'SPLICE_LENGTH_MID', 
                                       'SPLICE_LENGTH_RIGHT', 'PRESSURE_S1_L', 'PRESSURE_S1_R', 'PRESSURE_S2_L',
                                       'PRESSURE_S2_R', 'PRESSURE_S3_L', 'PRESSURE_S3_R', 'PRESSURE_S4_L',
                                       'PRESSURE_S4_R', 'PRESSURE_S5_L', 'PRESSURE_S5_R', 'PRESSURE_S6_L',
                                       'PRESSURE_S6_R', 'PRESSURE_S7_L', 'PRESSURE_S7_R', 'PRESSURE_S8_L', 
                                       'PRESSURE_S8_R'])
df_pa = df_pa.drop_duplicates(subset=['CUT_LENGTH', 'CONV_TIME', 'SPLICE_LENGTH_LEFT', 'SPLICE_LENGTH_MID', 
                                      'SPLICE_LENGTH_RIGHT', 'PRESSURE_S1_L', 'PRESSURE_S1_R', 'PRESSURE_S2_L',
                                      'PRESSURE_S2_R', 'PRESSURE_S3_L', 'PRESSURE_S3_R', 'PRESSURE_S4_L',
                                      'PRESSURE_S4_R', 'PRESSURE_S5_L', 'PRESSURE_S5_R', 'PRESSURE_S6_L',
                                      'PRESSURE_S6_R', 'PRESSURE_S7_L', 'PRESSURE_S7_R', 'PRESSURE_S8_L', 
                                      'PRESSURE_S8_R'])
# df_pa = df_pa.drop_duplicates()
print(df_pa.shape)

# Duplicate rows grouped by a column, typically ID column
# pd.concat(g for _, g in df_pa.groupby("SPLICE_OK") if len(g) > 1)

# df[ids.isin(ids[ids.duplicated()])].sort("ID")



Check missing data

df_pa['CUT_LENGTH'] = pd.to_numeric(df_pa['CUT_LENGTH'], errors='coerce')
df_pa['CONV_TIME'] = pd.to_numeric(df_pa['CONV_TIME'], errors='coerce')
df_pa['SPLICE_LENGTH_LEFT'] = pd.to_numeric(df_pa['SPLICE_LENGTH_LEFT'], errors='coerce')
df_pa['SPLICE_LENGTH_MID'] = pd.to_numeric(df_pa['SPLICE_LENGTH_MID'], errors='coerce')
df_pa['SPLICE_LENGTH_RIGHT'] = pd.to_numeric(df_pa['SPLICE_LENGTH_RIGHT'], errors='coerce')

df_pa['PRESSURE_S1_L'] = pd.to_numeric(df_pa['PRESSURE_S1_L'], errors='coerce')
df_pa['PRESSURE_S2_L'] = pd.to_numeric(df_pa['PRESSURE_S2_L'], errors='coerce')
df_pa['PRESSURE_S3_L'] = pd.to_numeric(df_pa['PRESSURE_S3_L'], errors='coerce')
df_pa['PRESSURE_S4_L'] = pd.to_numeric(df_pa['PRESSURE_S4_L'], errors='coerce')
df_pa['PRESSURE_S5_L'] = pd.to_numeric(df_pa['PRESSURE_S5_L'], errors='coerce')
df_pa['PRESSURE_S6_L'] = pd.to_numeric(df_pa['PRESSURE_S6_L'], errors='coerce')
df_pa['PRESSURE_S7_L'] = pd.to_numeric(df_pa['PRESSURE_S7_L'], errors='coerce')
df_pa['PRESSURE_S8_L'] = pd.to_numeric(df_pa['PRESSURE_S8_L'], errors='coerce')

df_pa['PRESSURE_S1_R'] = pd.to_numeric(df_pa['PRESSURE_S1_R'], errors='coerce')
df_pa['PRESSURE_S2_R'] = pd.to_numeric(df_pa['PRESSURE_S2_R'], errors='coerce')
df_pa['PRESSURE_S3_R'] = pd.to_numeric(df_pa['PRESSURE_S3_R'], errors='coerce')
df_pa['PRESSURE_S4_R'] = pd.to_numeric(df_pa['PRESSURE_S4_R'], errors='coerce')
df_pa['PRESSURE_S5_R'] = pd.to_numeric(df_pa['PRESSURE_S5_R'], errors='coerce')
df_pa['PRESSURE_S6_R'] = pd.to_numeric(df_pa['PRESSURE_S6_R'], errors='coerce')
df_pa['PRESSURE_S7_R'] = pd.to_numeric(df_pa['PRESSURE_S7_R'], errors='coerce')
df_pa['PRESSURE_S8_R'] = pd.to_numeric(df_pa['PRESSURE_S8_R'], errors='coerce')


total = df_pa.isnull().sum().sort_values(ascending=False)
percent = (df_pa.isnull().sum()/df_pa.isnull().count()*100).sort_values(ascending=False)
missing_data = pd.concat([total, percent], axis=1, keys=['Total', 'Percent'])
print(missing_data.head(10))
df_pa = df_pa.dropna()



total = df_pa.isnull().sum().sort_values(ascending=False)
percent = (df_pa.isnull().sum()/df_pa.isnull().count()*100).sort_values(ascending=False)
missing_data = pd.concat([total, percent], axis=1, keys=['Total', 'Percent'])
print(missing_data.head(20))



Adding ID column for slope calculations
Probably do not need it 

#df_pa.head
df_pa.to_csv('..\BridgestoneData\VMI_BP1_mod01.v1.csv', header=True, index=False)
