==> 04-FeatureEngineeringBP1-01.ipynb <==
Import python libraries required for the script

%matplotlib inline
%matplotlib notebook

from minepy import MINE
# from mpl_toolkits.mplot3d import Axes3D
# from numpy.polynomial import polynomial as npp
# from scipy import stats
# from scipy.stats import norm
from sklearn.ensemble import RandomForestClassifier
from sklearn.feature_selection import RFE
# from sklearn.feature_selection import RFECV
from sklearn.feature_selection import f_regression
# from sklearn.model_selection import StratifiedKFold
from sklearn.preprocessing import MinMaxScaler
# from sklearn.preprocessing import StandardScaler
# from sklearn.svm import SVC

# import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
# import seaborn as sns

# import warnings
# warnings.filterwarnings('ignore')

Read the CSV files

df = pd.read_csv('VMI_Data_BP1_V02_AK6_mod02.csv')

# Sampling after cleanup
# df_sample = df.sample(10000)

Check the columns in the data

print(df.columns)

df = df.drop(['LFT_SPLICE_LENGTH', 'MID_SPLICE_LENGTH', 'RHT_SPLICE_LENGTH', 'SPLICE_OK', 'LFT_SPLICE_GRADE', 
              'MID_SPLICE_GRADE', 'RHT_SPLICE_GRADE', 'LFT_SPLICE_DELTA', 'MID_SPLICE_DELTA', 'RHT_SPLICE_DELTA'
#               'LFT_SPLICE_MA5', 'LFT_SPLICE_MA10', 'LFT_SPLICE_MA20', 'LFT_SPLICE_MA50', 
#               'LFT_SPLICE_SLOPE5', 'LFT_SPLICE_SLOPE10', 'LFT_SPLICE_SLOPE20', 'LFT_SPLICE_SLOPE50',
#               , 'MID_SPLICE_MA5', 'MID_SPLICE_MA10', 'MID_SPLICE_MA20',
#               'MID_SPLICE_MA50', 'MID_SPLICE_SLOPE5', 'MID_SPLICE_SLOPE10', 'MID_SPLICE_SLOPE20',
#               'MID_SPLICE_SLOPE50',  'RHT_SPLICE_MA5', 'RHT_SPLICE_MA10', 
#               'RHT_SPLICE_MA20', 'RHT_SPLICE_MA50', 'RHT_SPLICE_SLOPE5', 'RHT_SPLICE_SLOPE10',
#               'RHT_SPLICE_SLOPE20', 'RHT_SPLICE_SLOPE50'
             ], axis=1)

print(df.columns)

print(df.shape)
df = df.drop_duplicates()

print(df.shape)

total = df.isnull().sum().sort_values(ascending=False)
percent = (df.isnull().sum()/df.isnull().count()*100).sort_values(ascending=False)
missing_data = pd.concat([total, percent], axis=1, keys=['Total', 'Percent'])
print(missing_data.head(10))
df = df.dropna()

def inverse(x):
    if x==0: 
        return 0
    return 1/x


df['CUT_LENGTH2'] = df['CUT_LENGTH']**2
df['CONV_WAIT_TIME2'] = df['CONV_WAIT_TIME']**2
df['BP1_S8_L2'] = df['BP1_S8_L']**2
df['BP1_S7_L2'] = df['BP1_S7_L']**2
df['BP1_S6_L2'] = df['BP1_S6_L']**2
df['BP1_S5_L2'] = df['BP1_S5_L']**2
df['BP1_S4_L2'] = df['BP1_S4_L']**2
df['BP1_S3_L2'] = df['BP1_S3_L']**2
df['BP1_S2_L2'] = df['BP1_S2_L']**2
df['BP1_S1_ML2'] = df['BP1_S1_ML']**2
df['BP1_S1_MR2'] = df['BP1_S1_MR']**2
df['BP1_S2_R2'] = df['BP1_S2_R']**2
df['BP1_S3_R2'] = df['BP1_S3_R']**2
df['BP1_S4_R2'] = df['BP1_S4_R']**2
df['BP1_S5_R2'] = df['BP1_S5_R']**2
df['BP1_S6_R2'] = df['BP1_S6_R']**2
df['BP1_S7_R2'] = df['BP1_S7_R']**2
df['BP1_S8_R2'] = df['BP1_S8_R']**2
df['LFT_SPLICE_PREV2'] = df['LFT_SPLICE_PREV']**2
df['MID_SPLICE_PREV2'] = df['MID_SPLICE_PREV']**2
df['RHT_SPLICE_PREV2'] = df['RHT_SPLICE_PREV']**2

df['CUT_LENGTH3'] = df['CUT_LENGTH'].apply(inverse)
df['CONV_WAIT_TIME3'] = df['CONV_WAIT_TIME'].apply(inverse)
df['BP1_S8_L3'] = df['BP1_S8_L'].apply(inverse)
df['BP1_S7_L3'] = df['BP1_S7_L'].apply(inverse)
df['BP1_S6_L3'] = df['BP1_S6_L'].apply(inverse)
df['BP1_S5_L3'] = df['BP1_S5_L'].apply(inverse)
df['BP1_S4_L3'] = df['BP1_S4_L'].apply(inverse)
df['BP1_S3_L3'] = df['BP1_S3_L'].apply(inverse)
df['BP1_S2_L3'] = df['BP1_S2_L'].apply(inverse)
df['BP1_S1_ML3'] = df['BP1_S1_ML'].apply(inverse)
df['BP1_S1_MR3'] = df['BP1_S1_MR'].apply(inverse)
df['BP1_S2_R3'] = df['BP1_S2_R'].apply(inverse)
df['BP1_S3_R3'] = df['BP1_S3_R'].apply(inverse)
df['BP1_S4_R3'] = df['BP1_S4_R'].apply(inverse)
df['BP1_S5_R3'] = df['BP1_S5_R'].apply(inverse)
df['BP1_S6_R3'] = df['BP1_S6_R'].apply(inverse)
df['BP1_S7_R3'] = df['BP1_S7_R'].apply(inverse)
df['BP1_S8_R3'] = df['BP1_S8_R'].apply(inverse)
df['LFT_SPLICE_PREV3'] = df['LFT_SPLICE_PREV'].apply(inverse)
df['MID_SPLICE_PREV3'] = df['MID_SPLICE_PREV'].apply(inverse)
df['RHT_SPLICE_PREV3'] = df['RHT_SPLICE_PREV'].apply(inverse)

scaler = MinMaxScaler()
mycols = ['CUT_LENGTH', 'CONV_WAIT_TIME', 'BP1_S8_L', 'BP1_S7_L', 'BP1_S6_L', 'BP1_S5_L', 'BP1_S4_L', 'BP1_S3_L', 
          'BP1_S2_L', 'BP1_S1_ML', 'BP1_S1_MR', 'BP1_S2_R', 'BP1_S3_R', 'BP1_S4_R', 'BP1_S5_R', 'BP1_S6_R',
          'BP1_S7_R', 'BP1_S8_R', 'PART_LENGTH', 'LFT_SPLICE_PREV', 'MID_SPLICE_PREV', 'RHT_SPLICE_PREV', 
          'CUT_LENGTH2', 'CONV_WAIT_TIME2', 'BP1_S8_L2', 'BP1_S7_L2', 'BP1_S6_L2', 'BP1_S5_L2', 
          'BP1_S4_L2', 'BP1_S3_L2', 'BP1_S2_L2', 'BP1_S1_ML2', 'BP1_S1_MR2', 'BP1_S2_R2', 'BP1_S3_R2', 'BP1_S4_R2', 
          'BP1_S5_R2', 'BP1_S6_R2',  'BP1_S7_R2', 'BP1_S8_R2', 'LFT_SPLICE_PREV2', 'MID_SPLICE_PREV2', 
          'RHT_SPLICE_PREV2','LFT_SPLICE_MA5', 'LFT_SPLICE_MA10', 'LFT_SPLICE_MA20', 'LFT_SPLICE_MA50',  
          'LFT_SPLICE_SLOPE5', 'LFT_SPLICE_SLOPE10', 'LFT_SPLICE_SLOPE20', 'LFT_SPLICE_SLOPE50', 
          'MID_SPLICE_MA5', 'MID_SPLICE_MA10', 'MID_SPLICE_MA20', 'MID_SPLICE_MA50', 'MID_SPLICE_SLOPE5', 
          'MID_SPLICE_SLOPE10', 'MID_SPLICE_SLOPE20', 'MID_SPLICE_SLOPE50', 'RHT_SPLICE_MA5', 
          'RHT_SPLICE_MA10','RHT_SPLICE_MA20', 'RHT_SPLICE_MA50', 'RHT_SPLICE_SLOPE5', 'RHT_SPLICE_SLOPE10',
          'RHT_SPLICE_SLOPE20', 'RHT_SPLICE_SLOPE50','CUT_LENGTH3', 'CONV_WAIT_TIME3', 'BP1_S8_L3', 'BP1_S7_L3',
          'BP1_S6_L3', 'BP1_S5_L3', 'BP1_S4_L3', 'BP1_S3_L3', 'BP1_S2_L3', 'BP1_S1_ML3', 'BP1_S1_MR3', 
          'BP1_S2_R3', 'BP1_S3_R3', 'BP1_S4_R3', 'BP1_S5_R3', 'BP1_S6_R3', 'BP1_S7_R3', 'BP1_S8_R3', 
          'LFT_SPLICE_PREV3', 'MID_SPLICE_PREV3', 'RHT_SPLICE_PREV3']
df[mycols] = scaler.fit_transform(df[mycols])

# df = df.groupby('SPLICE_GRADE', group_keys=False).apply(lambda x: x.sample(120))
# df['SPLICE_GRADE'] = df.apply(numericSG, axis=1) 

spliceGrade = {'Bad': 0, 'OK': 1, 'Good': 1} 
# spliceGrade = {'Bad': 0, 'OK': 1, 'Good': 2} 
df.SPLICE_GRADE = [spliceGrade[item] for item in df.SPLICE_GRADE] 


# df.to_csv('temp.csv', header=True, index=False)

X = df.drop(['SPLICE_GRADE'], axis=1)
y = df['SPLICE_GRADE']
y = y.astype('int')

def rank_to_dict(ranks, names, order=1):
    minmax = MinMaxScaler()
    ranks = minmax.fit_transform(order*np.array([ranks]).T).T[0]
    ranks = map(lambda x: round(x, 2), ranks)
    return dict(zip(names, ranks ))

np.random.seed(0)
 
size = 750

# names = ["x%s" % i for i in range(1,15)]
names = X.columns

ranks = {}

rf = RandomForestClassifier()
rf.fit(X,y)
ranks["RF"] = rank_to_dict(rf.feature_importances_, names)

#stop the search when 5 features are left (they will get equal scores)
rfe = RFE(rf, n_features_to_select=5)
rfe.fit(X,y)
ranks["RFE"] = rank_to_dict(list(map(float, rfe.ranking_)), names, order=-1)
 
f, pval  = f_regression(X, y, center=True)
ranks["Corr."] = rank_to_dict(f, names)
 
mine = MINE()
mic_scores = []
for i in range(X.shape[1]):
    mine.compute_score(X[:,i], y)
    m = mine.mic()
    mic_scores.append(m)
 
ranks["MIC"] = rank_to_dict(mic_scores, names) 
 
 
r = {}
for name in names:
    r[name] = round(np.mean([ranks[method][name] 
                             for method in ranks.keys()]), 2)
 
methods = sorted(ranks.keys())
ranks["Mean"] = r
methods.append("Mean")
 
print("\t%s" % "\t".join(methods))
for name in names:
    print("%s\t%s" % (name, "\t".join(map(str, 
                         [ranks[method][name] for method in methods]))))

ranks["RFE"] = rank_to_dict(list(map(float, rfe.ranking_)), names, order=-1)

f, pval  = f_regression(X, y, center=True)
ranks["Corr."] = rank_to_dict(f, names)
 
mine = MINE()
mic_scores = []
for i in range(X.shape[1]):
    mine.compute_score(X[:,i], y)
    m = mine.mic()
    mic_scores.append(m)
 
ranks["MIC"] = rank_to_dict(mic_scores, names) 
 
 
r = {}
for name in names:
    r[name] = round(np.mean([ranks[method][name] 
                             for method in ranks.keys()]), 2)
 
methods = sorted(ranks.keys())
ranks["Mean"] = r
methods.append("Mean")
 
print("\t%s" % "\t".join(methods))
for name in names:
    print("%s\t%s" % (name, "\t".join(map(str, 
                         [ranks[method][name] for method in methods]))))
